backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
pydantic-settings
axios
gunicorn
httpx
